change server id from long to integer

+++ web-server/src/main/java/com/abhimanyutech/servermanager/repository/ServerRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ServerRepository extends JpaRepository<Server, Long> {
+public interface ServerRepository extends JpaRepository<Server, Integer> {
     Server findByIpAddress(String ipAddress);
 }
